+++ resolwe/process/tests/test_python_process.py
@@ -11,6 +11,7 @@ from guardian.shortcuts import assign_perm
 from resolwe.flow.models import (
     Collection,
     Data,
+    DescriptorSchema,
     Entity,
     Process,
     Relation,
@@ -452,6 +453,39 @@ class PythonProcessTest(ProcessTestCase):
         self.assertEqual(len(data.process_error), 0)
         entity.refresh_from_db()
         self.assertEqual(entity.name, "New entity name")
+
+    @with_docker_executor
+    @tag_process("change-entity-descriptor")
+    def test_change_descriptor(self):
+        """Assign descriptor to entity."""
+
+        descriptor_schema = DescriptorSchema.objects.create(
+            name="Descriptor schema",
+            contributor=self.contributor,
+            schema=[
+                {
+                    "name": "Description",
+                    "type": "basic:string:",
+                    "default": "default value",
+                }
+            ],
+        )
+        entity = Entity.objects.create(
+            name="Entity", contributor=self.user, descriptor_schema=descriptor_schema
+        )
+
+        self.run_process(
+            "change-entity-descriptor",
+            {
+                "entity_id": entity.pk,
+                "description": "New description",
+            },
+        )
+        entity.refresh_from_db()
+        self.assertEqual(
+            entity.descriptor,
+            {"Description": "New description"},
+        )
 
 
 class PythonProcessRequirementsTest(ProcessTestCase):